Modernize Redux: AclThunks
Helps with #213.

Forgot to look at the thunks when modernizing
 aclSlice.  This rectifies that.

 Arguably some of these could have been
 turned into `createAsyncThunk` and have
 their results stored in redux. However, this
 would not only require major rewrites for
 the components using those thunks, but
 also child components who all rely on the
 current code structure. This was deemed
 too much work for what this PR wanted to
 achieve.
