+++ app/src/components/events/partials/ModalTabsAndPages/NewAccessPage.tsx
@@ -3,20 +3,20 @@ import { useTranslation } from "react-i18next";
 import cn from "classnames";
 import Notifications from "../../../shared/Notifications";
 import {
+	Role,
 	checkAcls,
 	fetchAclActions,
 	fetchAclTemplateById,
 	fetchAclTemplates,
 	fetchRolesWithTarget,
-} from "../../../../thunks/aclThunks";
+} from "../../../../slices/aclSlice";
 import { Field, FieldArray } from "formik";
-import { connect } from "react-redux";
 import RenderMultiField from "../../../shared/wizard/RenderMultiField";
 import { getUserInformation } from "../../../../selectors/userInfoSelectors";
 import { hasAccess } from "../../../../utils/utils";
 import DropDown from "../../../shared/DropDown";
 import { filterRoles, getAclTemplateText } from "../../../../utils/aclUtils";
-import { useAppSelector } from "../../../../store";
+import { useAppDispatch, useAppSelector } from "../../../../store";
 
 /**
  * This component renders the access page for new events and series in the wizards.
@@ -30,15 +30,14 @@ const NewAccessPage = ({
 	formik,
 // @ts-expect-error TS(7031): Binding element 'editAccessRole' implicitly has an... Remove this comment to see the full error message
 	editAccessRole,
-// @ts-expect-error TS(7031): Binding element 'checkAcls' implicitly has an 'any... Remove this comment to see the full error message
-	checkAcls,
 }) => {
 	const { t } = useTranslation();
+	const dispatch = useAppDispatch();
 
 	// States containing response from server concerning acl templates, actions and roles
-	const [aclTemplates, setAclTemplates] = useState([]);
-	const [aclActions, setAclActions] = useState([]);
-	const [roles, setRoles] = useState([]);
+	const [aclTemplates, setAclTemplates] = useState<{ id: string, value: string}[]>([]);
+	const [aclActions, setAclActions] = useState<{ id: string, value: string}[]>([]);
+	const [roles, setRoles] = useState<Role[]>([]);
 	const [loading, setLoading] = useState(false);
 
 	const user = useAppSelector(state => getUserInformation(state));
@@ -48,10 +47,8 @@ const NewAccessPage = ({
 		async function fetchData() {
 			setLoading(true);
 			const responseTemplates = await fetchAclTemplates();
-// @ts-expect-error TS(2345): Argument of type '{ id: string; value: any; }[]' i... Remove this comment to see the full error message
 			setAclTemplates(responseTemplates);
 			const responseActions = await fetchAclActions();
-// @ts-expect-error TS(2345): Argument of type '{ id: string; value: any; }[]' i... Remove this comment to see the full error message
 			setAclActions(responseActions);
 			const responseRoles = await fetchRolesWithTarget("ACL");
 			setRoles(responseRoles);
@@ -68,7 +65,7 @@ const NewAccessPage = ({
 
 		formik.setFieldValue("aclTemplate", value);
 		formik.setFieldValue("acls", template);
-		await checkAcls(formik.values.acls);
+		await dispatch(checkAcls(formik.values.acls));
 	};
 
 	return (
@@ -290,7 +287,7 @@ const NewAccessPage = ({
 																								write: false,
 																								actions: [],
 																							});
-																							checkAcls(formik.values.acls);
+																							dispatch(checkAcls(formik.values.acls));
 																						}}
                                             className="button-like-anchor"
 																					>
@@ -327,7 +324,7 @@ const NewAccessPage = ({
 					})}
 					disabled={!(formik.dirty && formik.isValid)}
 					onClick={async () => {
-						if (await checkAcls(formik.values.acls)) {
+						if (await dispatch(checkAcls(formik.values.acls))) {
 							nextPage(formik.values);
 						}
 					}}
@@ -349,16 +346,4 @@ const NewAccessPage = ({
 	);
 };
 
-// Getting state data out of redux store
-// @ts-expect-error TS(7006): Parameter 'state' implicitly has an 'any' type.
-const mapStateToProps = (state) => ({
-
-});
-
-// @ts-expect-error TS(7006): Parameter 'dispatch' implicitly has an 'any' type.
-const mapDispatchToProps = (dispatch) => ({
-// @ts-expect-error TS(7006): Parameter 'acls' implicitly has an 'any' type.
-	checkAcls: (acls) => dispatch(checkAcls(acls)),
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(NewAccessPage);
+export default NewAccessPage;
