+++ app/src/components/shared/wizard/WizardStepper.tsx
@@ -7,8 +7,8 @@ import {
 	useStepperStyle,
 } from "../../../utils/wizardUtils";
 import CustomStepIcon from "./CustomStepIcon";
-import { checkAcls } from "../../../thunks/aclThunks";
-import { connect } from "react-redux";
+import { checkAcls } from "../../../slices/aclSlice";
+import { useAppDispatch } from "../../../store";
 
 /**
  * This components renders the stepper navigation of new resource wizards
@@ -27,10 +27,9 @@ const WizardStepper = ({
 // @ts-expect-error TS(7031): Binding element 'setCompleted' implicitly has an '... Remove this comment to see the full error message
 	setCompleted,
 	hasAccessPage = false,
-// @ts-expect-error TS(7031): Binding element 'checkAcls' implicitly has an 'any... Remove this comment to see the full error message
-	checkAcls,
 }) => {
 	const { t } = useTranslation();
+	const dispatch = useAppDispatch();
 
 	const classes = useStepperStyle();
 
@@ -38,7 +37,7 @@ const WizardStepper = ({
 	const handleOnClick = async (key) => {
 		if (isSummaryReachable(key, steps, completed)) {
 			if (hasAccessPage) {
-				let check = await checkAcls(formik.values.acls);
+				let check = await dispatch(checkAcls(formik.values.acls));
 				if (!check) {
 					return;
 				}
@@ -80,10 +79,4 @@ const WizardStepper = ({
 	);
 };
 
-// @ts-expect-error TS(7006): Parameter 'dispatch' implicitly has an 'any' type.
-const mapDispatchToProps = (dispatch) => ({
-// @ts-expect-error TS(7006): Parameter 'acls' implicitly has an 'any' type.
-	checkAcls: (acls) => dispatch(checkAcls(acls)),
-});
-
-export default connect(null, mapDispatchToProps)(WizardStepper);
+export default WizardStepper;
