+++ app/src/components/users/partials/wizard/GroupRolesPage.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import WizardNavigationButtons from "../../../shared/wizard/WizardNavigationButtons";
 import SelectContainer from "../../../shared/wizard/SelectContainer";
-import { fetchRolesWithTarget } from "../../../../thunks/aclThunks";
+import { fetchRolesWithTarget } from "../../../../slices/aclSlice";
 
 /**
  * This component renders the role selection page of the new group wizard and group details modal
@@ -13,7 +13,7 @@ const GroupRolesPage = ({
     isEdit
 }: any) => {
 	// roles that can be chosen by user
-	const [roles, setRoles] = useState([]);
+	const [roles, setRoles] = useState<{ name: string }[]>([]);
 	// flag for API call
 	const [loading, setLoading] = useState(false);
 
@@ -30,7 +30,6 @@ const GroupRolesPage = ({
 					});
 				}
 			}
-// @ts-expect-error TS(2345): Argument of type '{ name: any; }[]' is not assigna... Remove this comment to see the full error message
 			setRoles(roleNames);
 			setLoading(false);
 		}
